public paths and shared libs are not so straightforward.
It seems that I need to know too much about how the remote bundle was built.

It seems that whatever the `publicPath` value of the remote build is, I need to host that bundle at exactly that URL, otherwise
all the module lookups that bundle will do will not work (they will look in the wrong place).

That feels pretty shit.